Replace `generate-temporaries' with portable implementation for codegen.
The values generated by `generate-temporaries' are not guaranteed to be
portable across Scheme implementations. As per the inline comments, the new
`gensym-values' implementation does not make as strong guarantees of
uniqueness, but should be sufficient for the purposes in which it is used
in this library.

Thanks to Takashi Kato <ktakashi@ymail.com>.

* protobuf/compile/codegen.scm (gensym-values): Replace
  `generate-temporaries' with a simple, counter-based implementation.